Split out `reference` module
